# Conflicts:
#	src/main/java/com/oopsw/erp_project/repository/ItemOrderDetailRepository.java
#	src/main/java/com/oopsw/erp_project/repository/ItemOrderRepository.java
#	src/main/java/com/oopsw/erp_project/repository/ItemRepository.java

+++ src/main/java/com/oopsw/erp_project/repository/ItemRepository.java
@@ -1,10 +1,11 @@
 package com.oopsw.erp_project.repository;
 
+import com.oopsw.erp_project.vo.Item;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
-public interface ItemRepository extends JpaRepository<Item, Long> {
+//@Repository // 써도 되고 안써도 되고
+public interface ItemRepository extends JpaRepository<Item, Integer> {
 
     List<Item> findByItemCode(String itemCode);
